requirements.txt
psutil==5.8.0
py-cpuinfo==8.0.0
